Frontend integration script new update

+++ Frontend Integration/Instantiate/socialMedia.ts
@@ -9,7 +9,7 @@ async function registerUser(name: string, bio: string, profileImageHash: string)
     const tx = await socialMediaInstance
         .connect(signer)
         .registerUser(name, bio, profileImageHash);
-    
+
     // Return the transaction object
     return tx;
 }
@@ -22,7 +22,7 @@ async function createPost(content: string, imgHash: string): Promise<ethers.Cont
     const tx = await socialMediaInstance
         .connect(signer)
         .createPost(content, imgHash);
-    
+
     // Return the transaction object
     return tx;
 }
@@ -31,11 +31,17 @@ async function createPost(content: string, imgHash: string): Promise<ethers.Cont
 async function upvotePost(postID: number): Promise<ethers.ContractTransaction> {
     const { signer, socialMediaInstance } = await getContract();
 
+    // Check if the user has already upvoted the post
+    const hasUpvoted = await socialMediaInstance.hasUpvoted(signer.address, postID);
+    if (hasUpvoted) {
+        throw new Error("You have already upvoted this post");
+    }
+
     // Call the upvote function of the contract with provided parameters
     const tx = await socialMediaInstance
         .connect(signer)
         .upvote(postID);
-    
+
     // Return the transaction object
     return tx;
 }
@@ -48,7 +54,7 @@ async function createComment(postId: number, content: string): Promise<ethers.Co
     const tx = await socialMediaInstance
         .connect(signer)
         .createComment(postId, content);
-    
+
     // Return the transaction object
     return tx;
 }
@@ -57,11 +63,17 @@ async function createComment(postId: number, content: string): Promise<ethers.Co
 async function deletePost(postId: number): Promise<ethers.ContractTransaction> {
     const { signer, socialMediaInstance } = await getContract();
 
+    // Check if the user is the original creator of the post
+    const originalCreator = await socialMediaInstance.getPostCreator(postId);
+    if (originalCreator !== signer.address) {
+        throw new Error("You are not authorized to delete this post");
+    }
+
     // Call the deletePost function of the contract with provided parameters
     const tx = await socialMediaInstance
         .connect(signer)
         .deletePost(postId);
-    
+
     // Return the transaction object
     return tx;
 }
@@ -70,11 +82,17 @@ async function deletePost(postId: number): Promise<ethers.ContractTransaction> {
 async function deleteComment(postId: number, commentId: number): Promise<ethers.ContractTransaction> {
     const { signer, socialMediaInstance } = await getContract();
 
+    // Check if the user is the original creator of the comment
+    const originalCreator = await socialMediaInstance.getCommentCreator(postId, commentId);
+    if (originalCreator !== signer.address) {
+        throw new Error("You are not authorized to delete this comment");
+    }
+
     // Call the deleteComment function of the contract with provided parameters
     const tx = await socialMediaInstance
         .connect(signer)
         .deleteComment(postId, commentId);
-    
+
     // Return the transaction object
     return tx;
 }
@@ -83,11 +101,17 @@ async function deleteComment(postId: number, commentId: number): Promise<ethers.
 async function editPost(postId: number, content: string, imgHash: string): Promise<ethers.ContractTransaction> {
     const { signer, socialMediaInstance } = await getContract();
 
+    // Check if the user is the original creator of the post
+    const originalCreator = await socialMediaInstance.getPostCreator(postId);
+    if (originalCreator !== signer.address) {
+        throw new Error("You are not authorized to edit this post");
+    }
+
     // Call the editPost function of the contract with provided parameters
     const tx = await socialMediaInstance
         .connect(signer)
         .editPost(postId, content, imgHash);
-    
+
     // Return the transaction object
     return tx;
 }
@@ -96,11 +120,17 @@ async function editPost(postId: number, content: string, imgHash: string): Promi
 async function editComment(postId: number, commentId: number, content: string): Promise<ethers.ContractTransaction> {
     const { signer, socialMediaInstance } = await getContract();
 
+    // Check if the user is the original creator of the comment
+    const originalCreator = await socialMediaInstance.getCommentCreator(postId, commentId);
+    if (originalCreator !== signer.address) {
+        throw new Error("You are not authorized to edit this comment");
+    }
+
     // Call the editComment function of the contract with provided parameters
     const tx = await socialMediaInstance
         .connect(signer)
         .editComment(postId, commentId, content);
-    
+
     // Return the transaction object
     return tx;
 }
@@ -109,11 +139,17 @@ async function editComment(postId: number, commentId: number, content: string): 
 async function downvote(postId: number): Promise<ethers.ContractTransaction> {
     const { signer, socialMediaInstance } = await getContract();
 
+    // Check if the user has already downvoted the post
+    const hasDownvoted = await socialMediaInstance.hasDownvoted(signer.address, postId);
+    if (hasDownvoted) {
+        throw new Error("You have already downvoted this post");
+    }
+
     // Call the downvote function of the contract with provided parameters
     const tx = await socialMediaInstance
         .connect(signer)
         .downvote(postId);
-    
+
     // Return the transaction object
     return tx;
 }
@@ -122,11 +158,17 @@ async function downvote(postId: number): Promise<ethers.ContractTransaction> {
 async function likeComment(postId: number, commentId: number): Promise<ethers.ContractTransaction> {
     const { signer, socialMediaInstance } = await getContract();
 
+    // Check if the user has already liked the comment
+    const hasLiked = await socialMediaInstance.hasLiked(signer.address, postId, commentId);
+    if (hasLiked) {
+        throw new Error("You have already liked this comment");
+    }
+
     // Call the likeComment function of the contract with provided parameters
     const tx = await socialMediaInstance
         .connect(signer)
         .likeComment(postId, commentId);
-    
+
     // Return the transaction object
     return tx;
 }
@@ -135,11 +177,17 @@ async function likeComment(postId: number, commentId: number): Promise<ethers.Co
 async function transferContractBalance(to: string): Promise<ethers.ContractTransaction> {
     const { signer, socialMediaInstance } = await getContract();
 
+    // Check if the signer has the authority to transfer the contract balance
+    const hasAuthority = await socialMediaInstance.hasAuthority(signer.address);
+    if (!hasAuthority) {
+        throw new Error("You don't have the authority to transfer the contract balance");
+    }
+
     // Call the transferContractBalance function of the contract with provided parameters
     const tx = await socialMediaInstance
         .connect(signer)
         .transferContractBalance(to);
-    
+
     // Return the transaction object
     return tx;
 }
@@ -150,7 +198,7 @@ async function getUserPosts(userId: number): Promise<any> {
 
     // Call the getUserPosts function of the contract with provided parameters
     const userPosts = await socialMediaInstance.getUserPosts(userId);
-    
+
     // Return the list of posts by the user
     return userPosts;
 }
@@ -161,7 +209,7 @@ async function getUpvotes(postId: number): Promise<any> {
 
     // Call the getUpvotes function of the contract with provided parameters
     const upvotes = await socialMediaInstance.getUpvotes(postId);
-    
+
     // Return the list of upvotes for the post
     return upvotes;
 }
@@ -172,7 +220,7 @@ async function getDownvotes(postId: number): Promise<any> {
 
     // Call the getDownvotes function of the contract with provided parameters
     const downvotes = await socialMediaInstance.getDownvotes(postId);
-    
+
     // Return the list of downvotes for the post
     return downvotes;
 }
@@ -183,7 +231,7 @@ async function getLikes(postId: number, commentId: number): Promise<any> {
 
     // Call the getLikes function of the contract with provided parameters
     const likes = await socialMediaInstance.getLikes(postId, commentId);
-    
+
     // Return the list of likes for the comment
     return likes;
 }
@@ -192,11 +240,16 @@ async function getLikes(postId: number, commentId: number): Promise<any> {
 async function followUser(userId: number): Promise<ethers.ContractTransaction> {
     const { signer, socialMediaInstance } = await getContract();
 
+    // Check if the user is trying to follow themselves
+    if (userId === signer.address) {
+        throw new Error("You cannot follow yourself");
+    }
+
     // Call the followUser function of the contract with provided parameters
     const tx = await socialMediaInstance
         .connect(signer)
         .followUser(userId);
-    
+
     // Return the transaction object
     return tx;
 }
@@ -209,7 +262,7 @@ async function unfollowUser(userId: number): Promise<ethers.ContractTransaction>
     const tx = await socialMediaInstance
         .connect(signer)
         .unfollowUser(userId);
-    
+
     // Return the transaction object
     return tx;
 }
